+++ test/clj/matcher_combinators/dispatch_test.clj
@@ -72,7 +72,8 @@
 
 (defn greater-than-matcher [expected-long]
   (matcher-combinators.core/->PredMatcher
-   (fn [actual] (> actual expected-long))))
+   (fn [actual] (> actual expected-long))
+   (str "greater than " expected-long)))
 
 (fact "example from docstring"
   5 => (match-with {java.lang.Long greater-than-matcher}