+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file. This
 change log follows the conventions of
 [keepachangelog.com](http://keepachangelog.com/).
 
+## [1.3.0]
+Add new matching context for clojure.test:
+- `match-roughly?`: Matches all numerics as long as they are within a given delta of the expected.
+
 ## [1.2.7]
 - Optionally allow omitting the first argument to `thrown-match?`: `(thrown-match? {:foo 1} (bang!)`